Introducing multifind.
Multifind takes a classifier and produces a list of examples that
are "diverse" with respect to that classifier. This is deliberately
a little underdefined, but right now basically means that it tries
to find a minimal example of each label.

The goal is to use this to rebuild given into being able to find
multiple exceptions at once, and then to allow example classification
on top of that. Minimal examples of each classification will be saved
in the database, leading to a more complete and consistent testing
experience across multiple runs.